backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
awscli==1.42.69
awsebcli==3.25.2
azure-core==1.36.0
azure-cosmos==4.14.1
bcrypt==4.0.1
blessed==1.23.0
botocore==1.40.69
cachetools==6.2.1
cement==2.10.14
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
docutils==0.19
dotenv==0.9.9
fabric==3.2.2
fastapi==0.115.4
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
invoke==2.2.1
jmespath==1.0.1
packaging==24.2
paramiko==4.0.0
passlib==1.7.4
pathspec==0.12.1
pip==25.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PyMuPDF==1.26.1
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
s3transfer==0.14.0
scikit-learn==1.7.2
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
termcolor==2.5.0
tinydb==4.8.2
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.32.0
vercel==0.3.5
vercel-sandbox==0.0.3
wcwidth==0.2.14
wheel==0.45.1
wrapt==2.0.1
